service-proxy/requirements.txt
aiohttp>=3.8.0  # Only needed if using async solution
fastapi>=0.68.0
uvicorn>=0.15.0
py-eureka-client>=0.10.0
requests>=2.26.0
python-dotenv>=0.19.0
pydantic-settings
python-jose[cryptography]==3.3.0
python-multipart>=0.0.5
PyJWT==2.4.0